chore: refactored scripts
